requirements.txt
beautifulsoup4==4.12.2
langchain==0.1.5
langchain-community==0.0.18
langchain-openai==0.0.5
pgvector==0.2.4
psycopg2-binary==2.9.9
pypdf==4.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
